requirements.txt
langchain
langchain-community
langchain-core
pymilvus
fastapi
uvicorn
python-dotenv
pydantic
sentence-transformers
python-multipart
langgraph
langchain_core
langchain_community
langchain_groq
